🎨:: 테마 관련 코드 구조 개선
