+++ .gitignore
@@ -64,5 +64,6 @@ google-services.json
 GoogleService-Info.plist
 
 # FVM (Flutter Version Management)
-.fvm/flutter_sdk
-.fvm/versions
+
+# FVM Version Cache
+.fvm/
