+++ lib/core/theme/config/dark_theme.dart
@@ -1,37 +1,35 @@
 import 'package:flutter/material.dart';
-import 'package:project_setting/core/theme/app_text_styles.dart';
-import 'app_colors.dart';
+import 'package:project_setting/core/theme/typography/app_text_styles.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
 
-/// 라이트 테마 설정
-class LightTheme {
-  LightTheme._();
+/// 다크 테마 설정
+class DarkTheme {
+  DarkTheme._();
 
   static ThemeData get theme => ThemeData(
     useMaterial3: true,
-    brightness: Brightness.light,
-    colorScheme: const ColorScheme.light(
+    brightness: Brightness.dark,
+    colorScheme: const ColorScheme.dark(
       primary: AppColors.mainColor,
-      surface: AppColors.bgSurface,
+      surface: AppColors.bgSurfaceDark,
       error: AppColors.negative,
     ),
-    scaffoldBackgroundColor: AppColors.background,
+    scaffoldBackgroundColor: AppColors.backgroundDark,
 
     // AppBar
     appBarTheme: AppBarTheme(
-      backgroundColor: AppColors.background,
+      backgroundColor: AppColors.backgroundDark,
       elevation: 0,
       centerTitle: false,
-      iconTheme: const IconThemeData(
-        color: AppColors.mainColor,
-      ),
+      iconTheme: const IconThemeData(color: AppColors.mainColor),
       titleTextStyle: AppTextStyles.title2.withColor(AppColors.mainColor),
     ),
 
     // Bottom Navigation Bar
     bottomNavigationBarTheme: const BottomNavigationBarThemeData(
-      backgroundColor: AppColors.bgSurface,
+      backgroundColor: AppColors.bgSurfaceDark,
       selectedItemColor: AppColors.sub2,
-      unselectedItemColor: AppColors.button,
+      unselectedItemColor: AppColors.sub2Dark,
       elevation: 0,
       type: BottomNavigationBarType.fixed,
       showSelectedLabels: false,
@@ -41,9 +39,9 @@ class LightTheme {
     // Input Decoration
     inputDecorationTheme: InputDecorationTheme(
       filled: true,
-      fillColor: AppColors.bgSurface,
-      labelStyle: AppTextStyles.text2.withColor(AppColors.mainText),
-      hintStyle: AppTextStyles.text3.withColor(AppColors.sub2),
+      fillColor: AppColors.bgSurfaceDark,
+      labelStyle: AppTextStyles.text2.withColor(AppColors.mainTextDark),
+      hintStyle: AppTextStyles.text3.withColor(AppColors.sub2Dark),
       contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 14),
       border: OutlineInputBorder(
         borderRadius: BorderRadius.circular(12),
