+++ lib/core/theme/config/light_theme.dart
@@ -1,37 +1,35 @@
 import 'package:flutter/material.dart';
-import 'package:project_setting/core/theme/app_text_styles.dart';
-import 'app_colors.dart';
+import 'package:project_setting/core/theme/typography/app_text_styles.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
 
-/// 다크 테마 설정
-class DarkTheme {
-  DarkTheme._();
+/// 라이트 테마 설정
+class LightTheme {
+  LightTheme._();
 
   static ThemeData get theme => ThemeData(
     useMaterial3: true,
-    brightness: Brightness.dark,
-    colorScheme: const ColorScheme.dark(
+    brightness: Brightness.light,
+    colorScheme: const ColorScheme.light(
       primary: AppColors.mainColor,
-      surface: AppColors.bgSurfaceDark,
+      surface: AppColors.bgSurface,
       error: AppColors.negative,
     ),
-    scaffoldBackgroundColor: AppColors.backgroundDark,
+    scaffoldBackgroundColor: AppColors.background,
 
     // AppBar
     appBarTheme: AppBarTheme(
-      backgroundColor: AppColors.backgroundDark,
+      backgroundColor: AppColors.background,
       elevation: 0,
       centerTitle: false,
-      iconTheme: const IconThemeData(
-        color: AppColors.mainColor,
-      ),
+      iconTheme: const IconThemeData(color: AppColors.mainColor),
       titleTextStyle: AppTextStyles.title2.withColor(AppColors.mainColor),
     ),
 
     // Bottom Navigation Bar
     bottomNavigationBarTheme: const BottomNavigationBarThemeData(
-      backgroundColor: AppColors.bgSurfaceDark,
+      backgroundColor: AppColors.bgSurface,
       selectedItemColor: AppColors.sub2,
-      unselectedItemColor: AppColors.sub2Dark,
+      unselectedItemColor: AppColors.button,
       elevation: 0,
       type: BottomNavigationBarType.fixed,
       showSelectedLabels: false,
@@ -41,9 +39,9 @@ class DarkTheme {
     // Input Decoration
     inputDecorationTheme: InputDecorationTheme(
       filled: true,
-      fillColor: AppColors.bgSurfaceDark,
-      labelStyle: AppTextStyles.text2.withColor(AppColors.mainTextDark),
-      hintStyle: AppTextStyles.text3.withColor(AppColors.sub2Dark),
+      fillColor: AppColors.bgSurface,
+      labelStyle: AppTextStyles.text2.withColor(AppColors.mainText),
+      hintStyle: AppTextStyles.text3.withColor(AppColors.sub2),
       contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 14),
       border: OutlineInputBorder(
         borderRadius: BorderRadius.circular(12),
