+++ lib/presentation/auth/login/model/login_state.dart
