+++ lib/presentation/auth/login/screens/login_screen.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:project_setting/core/constants/spacing_constants.dart';
-import 'package:project_setting/core/provider/auth_provider.dart';
 import 'package:project_setting/core/router/route_path.dart';
-import 'package:project_setting/core/theme/app_colors.dart';
-import 'package:project_setting/core/theme/app_text_styles.dart';
-import 'package:project_setting/presentation/auth/login/state/login_provider.dart';
-import 'package:project_setting/presentation/auth/login/state/login_state.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
+import 'package:project_setting/core/theme/typography/app_text_styles.dart';
+import 'package:project_setting/presentation/auth/auth_provider.dart';
+import 'package:project_setting/presentation/auth/login/model/login_state.dart';
+import 'package:project_setting/presentation/auth/login/viewModel/login_provider.dart';
 import 'package:project_setting/widgets/common/base_scaffold.dart';
 import 'package:project_setting/widgets/common/buttons/confirm_button.dart';
 import 'package:project_setting/widgets/common/textField/email_textField.dart';
