+++ lib/presentation/main_page/widget/outing_status.dart
@@ -1,4 +1,4 @@
-import 'package:project_setting/core/theme/app_colors.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
 import 'package:flutter/material.dart';
 
 enum OutingStatus {
