+++ lib/widgets/common/buttons/qr_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:project_setting/core/theme/app_colors.dart';
-import 'package:project_setting/core/theme/app_icons.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
+import 'package:project_setting/core/theme/icons/app_icons.dart';
 import 'package:project_setting/domain/enum/role_enum.dart';
 
 class QRButton extends StatelessWidget {
