+++ lib/widgets/common/textField/email_textField.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:project_setting/widgets/common/textField/base_textField.dart';
-import 'package:project_setting/core/theme/app_colors.dart';
-import 'package:project_setting/core/theme/app_text_styles.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
+import 'package:project_setting/core/theme/typography/app_text_styles.dart';
 
 /// 이메일 입력 텍스트 필드
 class EmailTextField extends StatelessWidget {
