+++ lib/widgets/common/textField/password_textField.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:project_setting/widgets/common/textField/base_textField.dart';
-import 'package:project_setting/core/theme/app_colors.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
 
 /// 비밀번호 입력 텍스트 필드 (보기/숨기기 기능 포함)
 class PasswordTextField extends StatefulWidget {
