+++ lib/widgets/common/textField/search_textField.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:project_setting/widgets/common/textField/base_textField.dart';
-import 'package:project_setting/core/theme/app_icons.dart';
-import 'package:project_setting/core/theme/app_colors.dart';
+import 'package:project_setting/core/theme/icons/app_icons.dart';
+import 'package:project_setting/core/theme/colors/app_colors.dart';
 
 /// 검색 텍스트 필드
 class SearchTextField extends StatefulWidget {
